fetch mycart data and added delete method

+++ src/Components/ChooseItem.jsx
@@ -1,42 +1,64 @@
-
-const ChooseItem = ({ item }) => {
+import { MdDeleteSweep } from 'react-icons/md';
+import Swal from 'sweetalert2';
+const ChooseItem = ({ item, myCartProduct, setMyCartProduct }) => {
 
     const { _id, name, brandName, type, price, rating, image } = item;
 
-    const handleDeleteToMyCart = id => {
-        console.log(id)
-        fetch(`http://localhost:5000/mycart/${id}`, {
-            method: "DELETE"
+    // const handleDeleteToMyCart = _id => {
+    //     console.log(_id)
+    //     fetch(`http://localhost:5000/mycart/${_id}`, {
+    //         method: "DELETE"
+    //     })
+    //         .then(res => res.json())
+    //         .then(data => {
+    //             console.log(data)
+                
+    //         })
+    // }
+     const handleDeleteToMyCart = (_id) => {
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                fetch(`http://localhost:5000/mycart/${_id}`, {
+                    method: "DELETE"
+                })
+                    .then(res => res.json())
+                    .then(data => {
+                        console.log(data)
+                        if (data.deletedCount > 0) {
+                            Swal.fire(
+                                'Deleted!',
+                                'Your Coffee has been deleted.',
+                                'success'
+                            )
+                            const remaining = myCartProduct.filter(item => item._id !== _id)
+                            setMyCartProduct(remaining);
+                        }
+                    })
+
+
+            }
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-            })
     }
 
     return (
-        <div className='flex gap-20 max-w-4xl mx-auto border rounded-lg  my-5 items-center p-5'>
+        <div className='flex gap-20 max-w-4xl bg-gray-50 mx-auto border rounded-lg  my-5  p-5'>
             <img className='w-36' src={image} alt="" />
 
-            <div className='text-left space-y-3'>
-                <h3 className="text-2xl">{name}</h3>
+            <div className='text-left space-y-2'>
+                <h3 className="text-2xl font-bold">{name}</h3>
                 <p>{brandName}</p>
-                <div className="flex gap-10">
-                    {/* <button className="border text-blue-700 rounded-lg py-2  px-5">{remote_or_onsite}</button>
-                <button className="border text-blue-700 rounded-lg py-2   px-5">{rating}</button> */}
-                </div>
-                <div className="flex gap-5">
-                    <div className='flex gap-2 items-center'>
-                        <p>{type}</p>
-                    </div>
-                    <div className='flex gap-2 items-center'>
-
-                        <p>Price : {price}</p>
-                    </div>
-                </div>
+                <p className="text-2xl font-bold">${price}</p>
             </div>
             <div>
-                <button onClick={() => handleDeleteToMyCart(_id)} className="btn btn-primary">Delete From Cart</button>
+                <div onClick={() => handleDeleteToMyCart(_id)} className="flex justify-center items-center gap-2 p-2 rounded-md bg-slate-100 hover:bg-slate-300 border cursor-pointer"> <MdDeleteSweep></MdDeleteSweep> <span>ARemove from cart</span></div>
             </div>
         </div>
     );
